dirtyvalue: Move to src/utility
