+++ src/displayapp/widgets/StatusIcons.h
@@ -6,6 +6,7 @@
 #include "components/battery/BatteryController.h"
 #include "components/ble/BleController.h"
 #include "displayapp/screens/BatteryIcon.h"
+#include "utility/DirtyValue.h"
 
 namespace Pinetime {
   namespace Applications {
@@ -27,10 +28,10 @@ namespace Pinetime {
         const Pinetime::Controllers::Battery& batteryController;
         const Controllers::Ble& bleController;
 
-        Screens::DirtyValue<uint8_t> batteryPercentRemaining {};
-        Screens::DirtyValue<bool> powerPresent {};
-        Screens::DirtyValue<bool> bleState {};
-        Screens::DirtyValue<bool> bleRadioEnabled {};
+        Utility::DirtyValue<uint8_t> batteryPercentRemaining {};
+        Utility::DirtyValue<bool> powerPresent {};
+        Utility::DirtyValue<bool> bleState {};
+        Utility::DirtyValue<bool> bleRadioEnabled {};
 
         lv_obj_t* bleIcon;
         lv_obj_t* batteryPlug;
